Remove 'labelled' field from assembly 'info' files
After much thought I realised that there is nothing to gain from labelling
assemblied with their asset code. The asset code assigned to an assembly is
more of a unique ID to differentiate it from other assemblies of the same
type rather than to match up a real life assembly with its inventory
counterpart.

The reason for grouping individual assets into assemblies is to allow the
relation of them over time to be evaluated. If we find that BeagleBoards
keep getting broken then we want to be able to see if it's always the same
power board that is breaking them.