Move BSG's now broken USB key into my possession
BSG's USB key broke.

BSG have a robot base with two motors on it.  Only one of the motor
controllers was connected on the SRIC bus (the other was connected to
the motor rail and its associated motor).  I observed this working a
couple of weeks ago when I was there.

The students there had done nothing with the robot over the last
couple of weeks.  The teacher had taken the robot home over half
term.  During half term he notified me that it had stopped working --
I asked for more information, but none came.

When I visited them it was revealed to me that the USB key did not
"appear" when it was plugged into a computer.  Plugging it into my
laptop showed that it did not enumerate :-(