Moved PSC2's memory stick into my posession
At the school last week, this memory stick was causing the
powerboard to boot and show the start screen very slowly --
upwards of 3 minutes towards the end of the meeting.

It was decided the memory stick was the problem for the
following reasons:

 - Using PSC1's hub, usb cable and memory stick with the
   same user code exported worked as expected, and quite
   quickly.

 - Using PSC2's hub and usb cable, and PSC1's memory stick
   with the same user code exported also worked quickly,
   as expected.

Interestingly, on plugging the memory stick into my PC, it
seems to work fine.  I haven't looked into it any further.
It hasn't been marked as broken as, for whatever reason,
it seems to work on my PC.