Fix a power board and flash it with the latest fw
I had to replace both the MSP430 and the LVDS driver, I suspect this
board received a huge amount of static at some point. I received the
board un-cased so this isn't that surprising.

The board has been flashed with the latest firmware and the BeagleBoard
has got the latest image on it.