Mark a IO board from GMR as broken.
Report states that SRIC will pass through the board; it doesn't appear
in the list of devices found by the power board though. Initial
suspicion is that the MSP isn't programmed for some reason.

GMR confirm that this is the first time they've tried to use it. It's
thus likely it was shipped in this state.