Move the component parts of Henry to my house to reassemble
The motive behind this was getting Henry ready for open days on the 4th and 5th of September.
I also took from the vault the drill, some cable ties, a belkin access point, screws, a bundle of short blue sric cables, a bundle of longer purple and pink sric cables, and a flat head screwdriver.