Mark a power board as broken.
The LCD on this assembly fails to display the SR boot screen; testing
with a known working LCD also fails, so it's either the BB or the power
board. Only thing that could be obviously wrong is the LCD socket not
being in properly, I'll re-seat it some other time.