Unknownify a motor board.
This appears to have bene marked broken with no additional data shortly
before shipping (SR2013). However I can't find anything wrong
electrically. It responds over SRIC, produces the right signals from the
motor output upon instruction, and successfully drivers a medium-size
motor. This might be able to be shipped after undergoing additional
testing.

It might be that this was marked broken as the case is broken.