Move a motor board between boxes.
I'm not in the vault right now, but we moved all the motor boards from
1RK2M into M1K2J, except one where it was difficult to identify the
board number. By the fact that this is the last remaining motor board in
that box, I can conclude that this was the board we couldn't read the
part code off.

(So, move it between the two boxes).