Add templates for the screw shields
The screw shields are in fact two parts.  There's one board for one
side of the headers on the Arduino, and one board for the other side.
I figured the easiest naming scheme to use for these sides was "power"
and "USB".  The power side is the one with the Arduino's power
jack on, and the USB side is the one with the Arduno's USB connector
on.