Swap a BB for a working one
The old one had a bad reset button and no u-boot. The new one has the
SR2012 image on it. This means that this power board assembly isn't
rated for the fedora-install-upgrade thing that was planned earlier in
the year. This is fine, because the competition is 10 days away, and
that update isn't going to happen.