Move all these things to storage
These are arranged in the storage locker along the right hand side
in two rows of boxes. At the far back corner are the SR history
boxes, next nearest outwards are the boxes most likely only needed
for the competition and nearest the front are things likely needed
for kickstart (including empty boxes).