Reinstate the proper name of this box
This was discovered by finding this box in storage, but not being
able to find it in the inventory. While I'm reasonably confident
that this is in the container, I'm not at all confident about its
contents so I'm leaving this in a location which is clearer that
it needs some investigation.